+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 
-* Updated to svd2rust 0.26.0, use tools binaries for CI
+* Replace `makehtml.py` with `svd2html`
+* Updated to svd2rust 0.28.0, svdtools 0.3.0, use tools binaries for CI
 * files in devices/common_patches moved to subdirectories
 
 ## [v0.15.1] 2022-07-04